requirements.txt
spotipy
requests
python-dotenv
tabulate
colorama
beautifulsoup4
playwright
streamlit